fix(database,webavanue): resolve duplicate indexes and build configuration
Database Module:
- Fixed type mismatch in SQLDelightContextPreferenceRepository.kt:118
  Changed decayFactor.toLong() to decayFactor.toDouble()
- Removed 12 duplicate SQL schema files from root directory
  Files moved to organized subdirectories (app/, element/, navigation/, stats/)
  Eliminates 25 duplicate index compilation errors

WebAvanue Module:
- Fixed project reference in universal/build.gradle.kts
  Changed :coredata to :Modules:WebAvanue:coredata
- Added 5 missing version catalog entries:
  * androidx-navigation (2.7.6)
  * compose-material-icons-extended
  * androidx-navigation-compose
  * androidx-security-crypto
  * sentry-android
- Fixed deprecated plugin references in webavanue/build.gradle.kts
- Fixed kotlin.test dependency reference

Result: All modules compile successfully without errors

+++ Modules/VoiceOS/core/database/src/commonMain/kotlin/com/augmentalis/database/repositories/impl/SQLDelightContextPreferenceRepository.kt
@@ -115,7 +115,7 @@ class SQLDelightContextPreferenceRepository(
         withContext(Dispatchers.Default) {
             database.transactionWithResult {
                 queries.applyTimeDecay(
-                    decayFactor = decayFactor.toLong(),
+                    decayFactor = decayFactor.toDouble(),
                     cutoffTime = currentTime
                 )
             }
